+++ src/Disclosure/index.js
@@ -74,7 +74,7 @@ export default class Disclosure extends AriaComponent {
     };
 
     // Merge remaining options with defaults and save all as instance properties.
-    Object.assign(this, { ...defaultOptions, ...options });
+    Object.assign(this, defaultOptions, options);
 
     // Initial component state.
     this.state = { expanded: this.loadOpen };
@@ -206,9 +206,6 @@ export default class Disclosure extends AriaComponent {
     } else {
       tabIndexDeny(this.interactiveChildElements);
     }
-
-    // Run {stateChangeCallback}
-    this.onStateChange.call(this, this.state);
   }
 
   /**