+++ src/Listbox/index.js
@@ -1,4 +1,3 @@
-import AriaComponent from '../AriaComponent';
 import Popup from '../Popup';
 import { setUniqueId } from '../lib/uniqueId';
 import keyCodes from '../lib/keyCodes';
@@ -10,7 +9,7 @@ import getFirstAndLastItems from '../lib/getFirstAndLastItems';
  *
  * https://www.w3.org/TR/wai-aria-practices-1.1/#Listbox
  */
-export default class ListBox extends AriaComponent {
+export default class ListBox extends Popup {
   /**
    * Create a ListBox.
    * @constructor
@@ -19,7 +18,8 @@ export default class ListBox extends AriaComponent {
    * @param {object}      options    The options object.
    */
   constructor(controller, options = {}) {
-    super(controller);
+    // Pass in the `listbox` type.
+    super(controller, { type: 'listbox' });
 
     /**
      * The string description for this object.
@@ -60,7 +60,7 @@ export default class ListBox extends AriaComponent {
     };
 
     // Merge remaining options with defaults and save all as instance properties.
-    Object.assign(this, { ...defaultOptions, ...options });
+    Object.assign(this, defaultOptions, options);
 
     // Bind class methods.
     this.preventWindowScroll = this.preventWindowScroll.bind(this);
@@ -69,11 +69,7 @@ export default class ListBox extends AriaComponent {
     this.targetHandleClick = this.targetHandleClick.bind(this);
     this.targetHandleBlur = this.targetHandleBlur.bind(this);
     this.scrollOptionIntoView = this.scrollOptionIntoView.bind(this);
-    this.onPopupStateChange = this.onPopupStateChange.bind(this);
-    this.show = this.show.bind(this);
-    this.hide = this.hide.bind(this);
     this.destroy = this.destroy.bind(this);
-    this.stateWasUpdated = this.stateWasUpdated.bind(this);
 
     this.init();
   }
@@ -82,6 +78,8 @@ export default class ListBox extends AriaComponent {
    * Set up the component's DOM attributes and event listeners.
    */
   init() {
+    super.init();
+
     /*
      * A reference to the class instance added to the controller and target
      * elements to enable external interactions with this instance.
@@ -122,22 +120,7 @@ export default class ListBox extends AriaComponent {
      *
      * @type {HTMLElement}
      */
-    this.state = { activeDescendant: this.firstOption };
-
-    /**
-     * The Listbox is basically a Popup to present a list of options, so we
-     * instantiate a Popup and subscribe to state changes to act on the Listbox
-     * when the Popup is shown and hidden.
-     *
-     * @type {Popup}
-     */
-    this.popup = new Popup(
-      this.controller,
-      {
-        type: 'listbox',
-        onStateChange: this.onPopupStateChange,
-      }
-    );
+    this.state.activeDescendant = this.firstOption;
 
     /*
      * Add the 'listbox' role to signify a component that presents a listbox of
@@ -154,7 +137,7 @@ export default class ListBox extends AriaComponent {
     this.target.setAttribute('tabindex', '-1');
 
     // Add event listeners.
-    this.controller.addEventListener('keydown', this.controllerHandleKeyup);
+    this.controller.addEventListener('keyup', this.controllerHandleKeyup);
     this.target.addEventListener('keydown', this.targetHandleKeydown);
     this.target.addEventListener('click', this.targetHandleClick);
     this.target.addEventListener('blur', this.targetHandleBlur);
@@ -170,11 +153,26 @@ export default class ListBox extends AriaComponent {
    * Track the selected Listbox option.
    * https://www.w3.org/TR/wai-aria-practices-1.1/#kbd_focus_activedescendant
    *
-   * @param {object} state The component state.
-   * @param {HTMLElement} state.activeDescendant The expected `activeDescendant` state.
+   * @param {string[]} updatedProps The newly-updated state properties.
    */
-  stateWasUpdated() {
+  stateWasUpdated(updatedProps) {
     const { activeDescendant, expanded } = this.state;
+
+    if (updatedProps.includes('expanded')) {
+      super.stateWasUpdated();
+
+      // The Popup is newly opened.
+      if (expanded) {
+        /*
+         * Focus the target (list) element when the Listbox is shown. Focus
+         * remains on the target element, with option selection coming through a
+         * combination of the `aria-selected` attribute on the option and the
+         * `aria-activedescendant` attribute on the target tracking the active
+         * option.
+         */
+        this.target.focus();
+      }
+    }
 
     if (expanded) {
       /*
@@ -188,41 +186,17 @@ export default class ListBox extends AriaComponent {
       activeDescendant.setAttribute('aria-selected', 'true');
 
       /*
-       * Track the newly selected option via the `aria-activedescendant` attribute
-       * on the target.
-       */
-      this.target.setAttribute('aria-activedescendant', activeDescendant.id);
-
-      /*
        * If the selected option is beyond the bounds of the list, scroll it into
        * view. Check this every time state is updated to ensure the selected
        * option is always visible.
        */
       this.scrollOptionIntoView(activeDescendant);
-    }
 
-    // Run {stateChangeCallback}
-    this.onStateChange.call(this, this.state);
-  }
-
-  /**
-   * Subscribe to Popup state changes.
-   *
-   * @param {object} popup.state the Popup state.
-   * @param {boolean} popup.state.expanded The Popup `expanded` state.
-   */
-  onPopupStateChange({ expanded }) {
-    const { activeDescendant } = this.state;
-
-    if (expanded) {
       /*
-       * Focus the target (list) element when the Listbox is shown. Focus
-       * remains on the target element, with option selection coming through a
-       * combination of the `aria-selected` attribute on the option and the
-       * `aria-activedescendant` attribute on the target tracking the active
-       * option.
+       * Track the newly selected option via the `aria-activedescendant`
+       * attribute on the target.
        */
-      this.target.focus();
+      this.target.setAttribute('aria-activedescendant', activeDescendant.id);
     } else {
       /*
        * When the Popup is hidden, the `aria-activedescendant` attribute should
@@ -241,9 +215,6 @@ export default class ListBox extends AriaComponent {
         this.controller.focus();
       }
     }
-
-    // Update component state.
-    this.setState({ activeDescendant, expanded });
   }
 
   /**
@@ -286,7 +257,6 @@ export default class ListBox extends AriaComponent {
     const { keyCode } = event;
     const {
       RETURN,
-      ESC,
       UP,
       DOWN,
       SPACE,
@@ -300,7 +270,7 @@ export default class ListBox extends AriaComponent {
        * need to update state here; if the Listbox is open rest assured an
        * option is selected.
        */
-      case ESC:
+      // ESC is handled via Popup.
       case RETURN:
       case SPACE: {
         event.preventDefault();
@@ -382,7 +352,7 @@ export default class ListBox extends AriaComponent {
    */
   targetHandleBlur() {
     // Use Popup state here, since the Popup drives the Listbox state.
-    if (this.popup.getState().expanded) {
+    if (this.state.expanded) {
       this.hide();
     }
   }
@@ -413,8 +383,8 @@ export default class ListBox extends AriaComponent {
    * Destroy the Listbox and Popup.
    */
   destroy() {
-    // Remove the references to the class instance.
-    this.deleteSelfReferences();
+    // Destroy the Popup.
+    super.destroy();
 
     // Remove the role attribute from each of the options.
     this.options.forEach((listItem) => {
@@ -426,9 +396,6 @@ export default class ListBox extends AriaComponent {
         listItem.removeAttribute('id');
       }
     });
-
-    // Destroy the Popup.
-    this.popup.destroy();
 
     // Remove the listbox role.
     this.target.removeAttribute('role');
@@ -444,19 +411,5 @@ export default class ListBox extends AriaComponent {
 
     // Run {destroyCallback}
     this.onDestroy.call(this);
-  }
-
-  /**
-   * Show the Listbox.
-   */
-  show() {
-    this.popup.show();
-  }
-
-  /**
-   * Hide the Listbox.
-   */
-  hide() {
-    this.popup.hide();
   }
 }