+++ src/Menu/index.js
@@ -88,7 +88,7 @@ export default class Menu extends AriaComponent {
     };
 
     // Merge remaining options with defaults and save all as instance properties.
-    Object.assign(this, { ...defaultOptions, ...options });
+    Object.assign(this, defaultOptions, options);
 
     // Bind class methods
     this.listHandleKeydown = this.listHandleKeydown.bind(this);