+++ src/MenuBar/MenuBar.test.js
@@ -71,7 +71,6 @@ const domElements = {
 const onInit = jest.fn();
 const onStateChange = jest.fn();
 const onDestroy = jest.fn();
-const onPopupInit = jest.fn();
 const { list } = domElements;
 
 const menuBar = new MenuBar(
@@ -81,7 +80,6 @@ const menuBar = new MenuBar(
     onInit,
     onStateChange,
     onDestroy,
-    onPopupInit,
   }
 );
 
@@ -92,10 +90,9 @@ describe('Menu collects DOM elements and adds attributes', () => {
     expect(domElements.list.menubar).toBeInstanceOf(MenuBar);
     expect(domElements.list.menubar.itemMatches).toEqual(':not(.exclude)');
 
-    expect(onInit).toHaveBeenCalled();
+    expect(onInit).toHaveBeenCalledTimes(1);
 
     expect(domElements.listThirdItem.popup).toBeInstanceOf(Popup);
-    expect(onPopupInit).toHaveBeenCalled();
   });
 
   it('Should add the correct DOM attributes and collect elements', () => {
@@ -127,7 +124,7 @@ describe('Menu correctly responds to events', () => {
       domElements.listFirstItem.focus();
       domElements.listFirstItem.dispatchEvent(keydownRight);
       expect(document.activeElement).toEqual(domElements.listSecondItem);
-      expect(onStateChange).toHaveBeenCalled();
+      expect(onStateChange).toHaveBeenCalledTimes(1);
     });
 
   it('Should move to the previous sibling list item with left arrow key',
@@ -135,6 +132,7 @@ describe('Menu correctly responds to events', () => {
       domElements.listSecondItem.focus();
       domElements.listSecondItem.dispatchEvent(keydownLeft);
       expect(document.activeElement).toEqual(domElements.listFirstItem);
+      expect(onStateChange).toHaveBeenCalledTimes(2);
     });
 
   it('Should move to the last list item with end key',
@@ -142,7 +140,7 @@ describe('Menu correctly responds to events', () => {
       domElements.listSecondItem.focus();
       domElements.listSecondItem.dispatchEvent(keydownEnd);
       expect(document.activeElement).toEqual(domElements.listLastItem);
-      expect(onStateChange).toHaveBeenCalled();
+      expect(onStateChange).toHaveBeenCalledTimes(3);
     });
 
   it('Should move to the first list item with home key',
@@ -214,7 +212,7 @@ describe('Menu correctly responds to events', () => {
     domElements.sublistOneSecondItem.addEventListener('click', onclick);
     domElements.sublistOneSecondItem.focus();
     domElements.sublistOneSecondItem.dispatchEvent(keydownSpace);
-    expect(onclick).toHaveBeenCalled();
+    expect(onclick).toHaveBeenCalledTimes(1);
   });
 });
 
@@ -235,7 +233,7 @@ describe('Menu should destroy properly', () => {
     expect(domElements.sublistTwoLastItem.getAttribute('tabindex')).toBeNull();
 
     expect(domElements.list.menubar).toBeUndefined();
-    expect(onDestroy).toHaveBeenCalled();
+    expect(onDestroy).toHaveBeenCalledTimes(1);
 
     // Quick and dirty verification that the original markup is restored.
     expect(document.body.innerHTML).toEqual(menubarMarkup);