+++ src/MenuButton/index.js
@@ -1,4 +1,3 @@
-import AriaComponent from '../AriaComponent';
 import Popup from '../Popup';
 import Menu from '../Menu';
 import keyCodes from '../lib/keyCodes';
@@ -9,16 +8,17 @@ import keyCodes from '../lib/keyCodes';
  * https://www.w3.org/TR/wai-aria-practices-1.1/#menubutton
  * https://www.w3.org/TR/wai-aria-practices-1.1/examples/menu-button/menu-button-links.html
  */
-export default class MenuButton extends AriaComponent {
+export default class MenuButton extends Popup {
   /**
-   * Create a ListBox.
+   * Create a MenuButton.
    * @constructor
    *
    * @param {HTMLElement} controller The activating element.
    * @param {object}      options    The options object.
    */
   constructor(controller, options = {}) {
-    super(controller);
+    // Pass in the `menu` type.
+    super(controller, { type: 'menu' });
 
     /**
      * The string description for this object.
@@ -26,9 +26,6 @@ export default class MenuButton extends AriaComponent {
      * @type {string}
      */
     this[Symbol.toStringTag] = 'MenuButton';
-
-    this.controller = controller;
-    this.target = super.constructor.getTargetElement(controller);
 
     /**
      * Options shape.
@@ -66,13 +63,10 @@ export default class MenuButton extends AriaComponent {
     };
 
     // Merge remaining options with defaults and save all as instance properties.
-    Object.assign(this, { ...defaultOptions, ...options });
+    Object.assign(this, defaultOptions, options);
 
     // Bind class methods.
     this.controllerHandleKeydown = this.controllerHandleKeydown.bind(this);
-    this.onPopupStateChange = this.onPopupStateChange.bind(this);
-    this.show = this.show.bind(this);
-    this.hide = this.hide.bind(this);
     this.destroy = this.destroy.bind(this);
 
     this.init();
@@ -82,22 +76,24 @@ export default class MenuButton extends AriaComponent {
    * Set up the component's DOM attributes and event listeners.
    */
   init() {
-    /**
-     * The MenuButton is basically a Popup to present a Menu, so we instantiate
-     * a Popup and subscribe to state changes to act on the MenuButton when the
-     * Popup is shown and hidden.
-     *
-     * @type {Popup}
-     */
-    this.popup = new Popup(
-      this.controller,
-      {
-        type: 'menu',
-        onStateChange: this.onPopupStateChange,
-      }
-    );
+    // Initialize Popup.
+    super.init();
 
-    // Initialize the Menu if we passed one in.
+    /*
+     * A reference to the class instance added to the controller and target
+     * elements to enable external interactions with this instance.
+     */
+    super.setSelfReference([this.controller, this.target]);
+
+    /**
+     * The MenuButton is a Popup to present a Menu. The element used as the Menu
+     * is determined by the following "logic":
+     * 1. An HTMLUListElement was passed in as the `list` option
+     * 2. The Popup target is an HTMLUListElement
+     * 3. The first HTMLUListElement we inside the target
+     *
+     * @type {Menu}
+     */
     if (null != this.list && 'UL' === this.list.nodeName) {
       this.menu = new Menu(this.list);
     } else if ('UL' === this.target.nodeName) {
@@ -111,31 +107,12 @@ export default class MenuButton extends AriaComponent {
     // Additional event listener(s).
     this.controller.addEventListener('keydown', this.controllerHandleKeydown);
 
-    /**
-     * Set initial state.
-     *
-     * @type {object}
-     */
-    this.state = { expanded: false };
-
     // Run {initCallback}
     this.onInit.call(this);
   }
 
   /**
-   * Keep this component's state synced with the Popup's state.
-   *
-   * @param {Object} state The Popup state.
-   */
-  onPopupStateChange(state) {
-    this.setState(state);
-
-    // Run {stateChangeCallback}
-    this.onStateChange.call(this, this.state);
-  }
-
-  /**
-   * Handle keydown events on the MenuButton controller.
+   * Handle additional keydown events on the MenuButton controller.
    *
    * @param {Event} event The event object.
    */
@@ -148,41 +125,32 @@ export default class MenuButton extends AriaComponent {
       SPACE,
     } = keyCodes;
 
-    switch (keyCode) {
-      /*
-       * Open the menu and move focus to the first menu item.
-       */
-      case RETURN:
-      case SPACE:
-      case DOWN: {
-        event.preventDefault();
-        this.show();
+    if ([RETURN, SPACE, UP, DOWN].includes(keyCode)) {
+      event.preventDefault();
 
-        // Move focus to the first menu item.
-        if (this.menu.firstItem) {
+      // RETURN and SPACE are handled in the parent class.
+      if ([UP, DOWN].includes(keyCode)) {
+        this.toggle();
+      }
+
+      // Get fresh state.
+      const { expanded } = this.state;
+
+      /*
+       * UP moves to last menu item, the rest move to first menu item.
+       */
+      if (expanded) {
+        if (
+          [RETURN, SPACE, DOWN].includes(keyCode)
+          && null != this.menu.firstItem
+        ) {
           this.menu.firstItem.focus();
         }
 
-        break;
-      }
-
-      /*
-       * Opens the menu and move focus to the last menu item.
-       */
-      case UP: {
-        event.preventDefault();
-        this.show();
-
-        // Move focus to the last menu item.
-        if (this.menu.lastItem) {
+        if (keyCode === UP && null != this.menu.lastItem) {
           this.menu.lastItem.focus();
         }
-
-        break;
       }
-
-      default:
-        break;
     }
   }
 
@@ -190,8 +158,8 @@ export default class MenuButton extends AriaComponent {
    * Destroy the Popup and Menu.
    */
   destroy() {
-    // Destroy the MenuButton Popup.
-    this.popup.destroy();
+    // Destroy the Popup.
+    super.destroy();
 
     // Destroy the Menu.
     this.menu.destroy();
@@ -204,19 +172,5 @@ export default class MenuButton extends AriaComponent {
 
     // Run {destroyCallback}
     this.onDestroy.call(this);
-  }
-
-  /**
-   * Show the menu Popup.
-   */
-  show() {
-    this.popup.show();
-  }
-
-  /**
-   * Hide the menu Popup.
-   */
-  hide() {
-    this.popup.hide();
   }
 }